IHashAlgoInfo 接口增加 DigestLength 属性

+++ HashCalculator/HashAlgos/OfficialImplBlake2b.cs
@@ -23,13 +23,13 @@ namespace HashCalculator
         [DllImport(Embedded.HashAlgs, CallingConvention = CallingConvention.Cdecl)]
         private static extern int blake2b_final(IntPtr statePtr, byte[] output, ulong outlen);
 
-        public override ulong MaxOutputSize => 64;
+        public override int MaxOutputSize => 64;
 
         public override string NamePrefix => "BLAKE2b";
 
         public override AlgoType AlgoGroup => AlgoType.BLAKE2B;
 
-        public OfficialImplBlake2b(ulong bitLength) : base(bitLength)
+        public OfficialImplBlake2b(int bitLength) : base(bitLength)
         {
         }
